moved chips to a component file and added chip data model
