+++ lib/model/events_M/chip_model.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+
+class ChipData {
+  final String label;
+  final Icon icon;
+
+  ChipData(this.label, this.icon);
+}
